update controls page and idp page to support personal device enrollment (#31857)
relates to #30662

updates the controls end user auth section to show the end user
experience for personally enrolled devices

<img width="1608" height="910" alt="image"
src="https://github.com/user-attachments/assets/cb3ef3d8-21d3-49dd-9af2-837cba7ebd00"
/>

also moves the end user auth section that was in mdm to the idp setting
page

<img width="1161" height="1010" alt="image"
src="https://github.com/user-attachments/assets/df1bb386-acb7-4b99-a714-4bc7c7e9f8fc"
/>

also has a small fix to add to show the user cards for ios and iPad
devices now


- [x] Changes file added for user-visible changes in `changes/`,
`orbit/changes/` or `ee/fleetd-chrome/changes`.
- [ ] Added/updated automated tests
- [x] QA'd all new/changed functionality manually
