+++ frontend/pages/ManageControlsPage/SetupExperience/cards/EndUserAuthentication/components/EndUserAuthForm/EndUserAuthForm.tsx
@@ -64,8 +64,10 @@ const EndUserAuthForm = ({
         <p className={classes}>
           Require end users to authenticate with your identity provider (IdP)
           and agree to an end user license agreement (EULA) when they setup
-          their new macOS hosts.{" "}
-          <Link to={PATHS.ADMIN_INTEGRATIONS_MDM}>View IdP and EULA</Link>
+          their new macOS, iOS, iPadOS and Android hosts.{" "}
+          <Link to={PATHS.ADMIN_INTEGRATIONS_IDENTITY_PROVIDER}>
+            View IdP and EULA
+          </Link>
         </p>
         <GitOpsModeTooltipWrapper
           renderChildren={(disableChildren) => (
