+++ frontend/pages/admin/IntegrationsPage/cards/IdentityProviders/_styles.scss
@@ -1,15 +1,5 @@
 .identity-providers {
-
-  &__section-card-content {
-    margin: 0;
-  }
-
-  &__received-tooltip {
-    display: inline-flex;
-  }
-
-  &__learn-more-link {
-    // used to keep the "Learn more" text content on a single line.
-    min-width: 88px;
-  }
+    display: flex;
+    flex-direction: column;
+    gap: $pad-xxlarge;
 }
