+++ frontend/pages/admin/IntegrationsPage/cards/IdentityProviders/components/IdentityProviderSection/_styles.scss
@@ -1,0 +1,14 @@
+.identity-provider-section {
+  &__section-card-content {
+    margin: 0;
+  }
+
+  &__received-tooltip {
+    display: inline-flex;
+  }
+
+  &__learn-more-link {
+    // used to keep the "Learn more" text content on a single line.
+    min-width: 88px;
+  }
+}
