+++ frontend/pages/admin/IntegrationsPage/cards/IdentityProviders/components/IdentityProviderSection/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./IdentityProviderSection";
