+++ frontend/pages/admin/IntegrationsPage/cards/MdmSettings/MdmSettings.tsx
@@ -13,7 +13,6 @@ import mdmAPI, { IEulaMetadataResponse } from "services/entities/mdm";
 
 import MdmSettingsSection from "./components/MdmSettingsSection";
 import AppleBusinessManagerSection from "./components/AppleBusinessManagerSection";
-import IdpSection from "./components/IdpSection";
 import EulaSection from "./components/EulaSection";
 import EndUserMigrationSection from "./components/EndUserMigrationSection";
 import WindowsAutopilotSection from "./components/WindowsAutopilotSection";
@@ -132,7 +131,6 @@ const MdmSettings = ({ router }: IMdmSettingsProps) => {
           />
           {isPremiumTier && !!config?.mdm.apple_bm_enabled_and_configured && (
             <>
-              <IdpSection />
               <EulaSection
                 eulaMetadata={eulaMetadata}
                 isEulaUploaded={!noEulaUploaded}
