--- frontend/pages/admin/IntegrationsPage/cards/MdmSettings/components/IdpSection/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./IdpSection";
